fix(analyzer): Change default get behaviour
