--- core/opentaint-jvm-sast-dataflow/src/main/kotlin/org/opentaint/jvm/sast/dataflow/JIRMethodGetDefaultProvider.kt
@@ -1,62 +1,0 @@
-package org.opentaint.jvm.sast.dataflow
-
-import org.opentaint.dataflow.ap.ifds.access.FactAp
-import org.opentaint.dataflow.configuration.jvm.CopyAllMarks
-import org.opentaint.dataflow.configuration.jvm.PositionAccessor
-import org.opentaint.dataflow.configuration.jvm.PositionWithAccess
-import org.opentaint.dataflow.configuration.jvm.Result
-import org.opentaint.dataflow.configuration.jvm.TaintPassThrough
-import org.opentaint.dataflow.configuration.jvm.This
-import org.opentaint.dataflow.configuration.mkTrue
-import org.opentaint.dataflow.jvm.ap.ifds.taint.TaintRulesProvider
-import org.opentaint.ir.api.common.CommonMethod
-import org.opentaint.ir.api.common.cfg.CommonInst
-import org.opentaint.ir.api.jvm.JIRMethod
-import org.opentaint.ir.api.jvm.TypeName
-import org.opentaint.ir.impl.cfg.util.isArray
-import org.opentaint.ir.impl.types.TypeNameImpl
-
-class JIRMethodGetDefaultProvider(
-    val base: TaintRulesProvider,
-    private val projectClasses: ClassLocationChecker,
-) : TaintRulesProvider by base {
-    override fun passTroughRulesForMethod(
-        method: CommonMethod,
-        statement: CommonInst?,
-        fact: FactAp?,
-        allRelevant: Boolean
-    ): Iterable<TaintPassThrough> {
-        val baseRules = base.passTroughRulesForMethod(method, statement, fact, allRelevant)
-
-        // note: hack to disable default get provide in Alias Analysis
-        if (statement == null) return baseRules
-
-        if (method !is JIRMethod || method.isStatic) return baseRules
-
-        if (!method.name.startsWith("get")) return baseRules
-
-        if (projectClasses.isProjectClass(method.enclosingClass)) return baseRules
-
-        var actions = getDefaultActions
-        if (method.returnType.mayBeArray()) {
-            actions = actions + getDefaultArrayActions
-        }
-
-        val getDefaultRule = TaintPassThrough(method, mkTrue(), actions, info = null)
-        return baseRules + getDefaultRule
-    }
-
-    companion object {
-        private val objectTypeName = TypeNameImpl.fromTypeName("java.lang.Object")
-
-        private fun TypeName.mayBeArray(): Boolean = isArray || this == objectTypeName
-
-        private val getDefaultActions = listOf(
-            CopyAllMarks(from = This, to = Result)
-        )
-
-        private val getDefaultArrayActions = listOf(
-            CopyAllMarks(from = This, to = PositionWithAccess(Result, PositionAccessor.ElementAccessor))
-        )
-    }
-}
